Error Stack Experiment (#1099)
* [C] Initial test and append to stack.

* [C] Working error stack.

* [C] Implement APPEND error

* [C] Integrate error stack into aeron_udp_channel_transport_init.  Remove some duplication of error descriptions.

* [C] Covert lots of code over to AERON_SET_ERROR and AERON_APPEND_ERROR.

* [C] Add test for various error cases in the driver and client.  Make
generic error response non-zero.

* [C] Simplify use of distinct error log.  I.e there is just a single string passed to the logger and we match against all of it.  Has closer matching behaviour to Java implemention.  Revise and simplify use of distinct error log in Sender, Receiver, Conductor and Name Re-resolution.  More tests.

* [C] Use a single buffer for the error state instead of an error stack.  More efficient use of memory and less thought required for nesting errors returned from the server.  Handle truncation of error state with simple ellipsis.

* [C] Fix segfault.

* [C] Don't use system dependent name in test.

* [C] Always cast size_t to uint64_t for printf style formatting.

* [C] Remove formatted args.

* [C] Have verification of messages in distinct log retry and timeout.  Look across all observations for a match.

* [C] Fix warning.

* [C] Extra checking to see that the cnc file has been closed correctly.

+++ aeron-samples/src/main/c/basic_publisher.c
@@ -184,11 +184,11 @@ int main(int argc, char **argv)
     for (size_t i = 0; i < messages && is_running(); i++)
     {
 #if defined(_MSC_VER)
-        const int message_len = sprintf_s(message, sizeof(message) - 1, "Hello World! %" PRIu32, (uint32_t)i);
+        const int message_len = sprintf_s(message, sizeof(message) - 1, "Hello World! %" PRIu64, (uint64_t)i);
 #else
-        const int message_len = snprintf(message, sizeof(message) - 1, "Hello World! %" PRIu32, (uint32_t)i);
+        const int message_len = snprintf(message, sizeof(message) - 1, "Hello World! %" PRIu64, (uint64_t)i);
 #endif
-        printf("offering %" PRIu32 "/%" PRIu32 " - ", (uint32_t)i, (uint32_t)messages);
+        printf("offering %" PRIu64 "/%" PRIu64 " - ", (uint64_t)i, (uint64_t)messages);
         fflush(stdout);
 
         int64_t result = aeron_publication_offer(
